+++ MOP/src/Hypervisor.cs
@@ -101,11 +101,6 @@ namespace MOP
             cursorFSM = GameObject.Find("PLAYER").GetPlayMakerByName("Update Cursor");
             cursorFSM.enabled = false;
 
-            // Start the delayed initialization routine
-            PlayMakerFSM[] gtGrille = GameObject.Find("grille gt(Clone)").GetComponents<PlayMakerFSM>();
-            foreach (var fsm in gtGrille)
-                fsm.Fsm.RestartOnEnable = false;
-
             // Disable rule files if user wants it.
             if (!RulesManager.Instance.LoadRules)
             {
@@ -119,7 +114,8 @@ namespace MOP
         #region MOP Initialization
         IEnumerator DelayedInitializaitonRoutine()
         {
-            yield return new WaitForSeconds(2);
+            //yield return new WaitForSeconds(2);
+            yield return null;
             Initialize();
         }
 
@@ -171,7 +167,6 @@ namespace MOP
                 ExceptionManager.New(ex, false, "WORLD_OBJECTS_1_INITIALIZAITON_FAIL");
             }
 
-
             // Initialize places.
             placeManager = new PlaceManager();
 
@@ -205,7 +200,6 @@ namespace MOP
                 worldObjectManager.Add("SwampColliders", DisableOn.PlayerInHome);
                 worldObjectManager.Add("RYKIPOHJA", DisableOn.PlayerInHome);
                 worldObjectManager.Add("COMPUTER", DisableOn.PlayerAwayFromHome);
-                //worldObjectManager.Add("JOBS/HouseDrunkNew", DisableOn.PlayerInHome);
 
                 ModConsole.Print("[MOP] World objects (2) loaded");
             }